split assignment.py tests
